+++ packages/core/test/unit/impl/RetryBuffer.test.ts
@@ -25,22 +25,58 @@ describe('RetryBuffer', () => {
     }
     await waitForCondition(() => output.length >= 10)
     expect(output).length.is.greaterThan(0)
-    subject.addLines([], 1, 1, Date.now() + 1000) // shall be ignored
+    subject.addLines([], 1, 1, Date.now() + 1000) // empty lines are ignored
     subject.close()
-    subject.addLines(['x'], 1, 1, Date.now() + 1000) // shall be ignored
+    subject.addLines(['x'], 1, 1, Date.now() + 1000) // ignored after calling flush
     await subject.flush()
     expect(input).deep.equals(output)
   })
   it('ignores lines on heavy load', async () => {
-    const input = [] as Array<[string[], number]>
     const output = [] as Array<[string[], number]>
-    const subject = new RetryBuffer(5, (lines, countdown) => {
-      output.push([lines, countdown])
-      return Promise.resolve()
-    })
+    const subject = new RetryBuffer(
+      5,
+      (lines, countdown) => {
+        output.push([lines, countdown])
+        return Promise.resolve()
+      },
+      ({expires}) => {
+        for (let actual = subject.first; actual; actual = actual?.next) {
+          if (actual.expires < expires) {
+            expect.fail(
+              `${actual.expires} entry was not removed from the retry buffer, but ${expires} was!`
+            )
+          }
+        }
+      }
+    )
     for (let i = 0; i < 10; i++) {
-      if (i >= 5) input.push([['a' + i], i])
-      subject.addLines(['a' + i], i, 100, Date.now() + 1000)
+      subject.addLines(['a' + i], i, 100 + i, Date.now() + 1000 + i)
+    }
+    await subject.flush()
+    expect(subject.close()).equals(0)
+    expect(logs.error).length.is.greaterThan(0) // 5 entries over limit
+    expect(output).length.is.lessThan(6) // at most 5 items will be written
+  })
+  it('ignores lines on heavy load with descending expiration', async () => {
+    const output = [] as Array<[string[], number]>
+    const subject = new RetryBuffer(
+      5,
+      (lines, countdown) => {
+        output.push([lines, countdown])
+        return Promise.resolve()
+      },
+      ({expires}) => {
+        for (let actual = subject.first; actual; actual = actual?.next) {
+          if (actual.expires < expires) {
+            expect.fail(
+              `${actual.expires} entry was not removed from the retry buffer, but ${expires} was!`
+            )
+          }
+        }
+      }
+    )
+    for (let i = 0; i < 10; i++) {
+      subject.addLines(['a' + i], i, 100 + i, Date.now() + 1000 - i)
     }
     await subject.flush()
     expect(subject.close()).equals(0)